♻️refactor: 자동차 경기 출력 부분 클래스로 분리
